refactor(ledger): extract sqlite operations
